% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fibonacci.R
\name{fibonacci}
\alias{fibonacci}
\title{Fibonacci Function}
\usage{
fibonacci(n)
}
\arguments{
\item{n}{a number that represent the number of values you want returned}
}
\value{
The output is a printed sequence of numbers. This is the fibonacci sequence up to n.
}
\description{
Return the first n values in the fibonacci sequence.
}
\details{
n can not be negative or 0. The function takes n and creates a loop that for each value from 1 to n the fibonacci sequence is calculated and the next number is printed. The loop ends when n is reached.
}
\examples{
fibonacci(2)
fibonacci(4)
}
