% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnAreaCircle.R
\name{fnAreaCircle}
\alias{fnAreaCircle}
\title{Calculate the area of a circle given its radius}
\usage{
fnAreaCircle(r)
}
\arguments{
\item{r}{radius}
}
\value{
area circle
}
\description{
Calculate the area of a circle given its radius
}
